Changes to hashable file watcher.
The previous logic was a bit off with regard to restarts. Because the current
modified info of files was being saved between restarts, this is the info with
respect to which modifications were being detected. This means upon every
restart, the FileWatcher would not report that there were any differences.
Instead, the visibility state of the "connection table needs recompiling"
notification was being saved across restarts, such that it would display if it
was being displayed previously, regardless of the file status. This means that
with the hashing file watcher, a file being restored to its previous state would
not make the notification go away.

The logic has therefore been changed to store the 'clean' info of the files
across restarts, and updating to the current modified info of the files only
upon recompilation+restart, or when the files are first added to be watched.

Have the callback respond to deleted files as well.