Move new_rep_name to runmanager, since the runmanager compilation queue uses it too now.
Inform runmanager how many shots are remaining each iteration of the queue.
This is a bit of a hack presently, it happens in the queue's thread
and so blocks for up to 5 seconds if runmanager is not running.

It would be better done as an extension that the queue pipes events to.